+++ lib/src/widgets/skadi_confirmation_dialog.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
+import '../utilities/types.dart';
 import 'skadi_platform_checker.dart';
 
 ///Create a dialog that provide onCancel and Confirm action
@@ -9,9 +10,12 @@ class SkadiConfirmationDialog extends StatelessWidget {
   final String title;
   final String cancelText;
   final String confirmText;
-  final VoidCallback? onCancel;
-  final VoidCallback? onConfirm;
+  final FutureOrCallBack? onCancel;
+  final FutureOrCallBack? onConfirm;
+  //
   final ShapeBorder? androidDialogShape;
+  final Color? confirmTextColor;
+  final Color? cancelTextColor;
 
   ///Mark confirm action as a danger action
   final bool _danger;
@@ -26,6 +30,8 @@ class SkadiConfirmationDialog extends StatelessWidget {
     this.onCancel,
     this.onConfirm,
     this.androidDialogShape,
+    this.confirmTextColor,
+    this.cancelTextColor,
   })  : _danger = false,
         super(key: key);
 
@@ -38,6 +44,8 @@ class SkadiConfirmationDialog extends StatelessWidget {
     this.onCancel,
     this.onConfirm,
     this.androidDialogShape,
+    this.confirmTextColor,
+    this.cancelTextColor,
   })  : _danger = true,
         super(key: key);
 
@@ -54,18 +62,24 @@ class SkadiConfirmationDialog extends StatelessWidget {
       CupertinoDialogAction(
         isDestructiveAction: !_danger,
         onPressed: () {
-          onCancel?.call();
-          Navigator.of(context).pop(false);
+          if (onCancel != null) {
+            onCancel!.call();
+          } else {
+            Navigator.pop(context, false);
+          }
         },
         child: Text(cancelText),
       ),
       CupertinoDialogAction(
         isDestructiveAction: _danger,
-        child: Text(confirmText),
         onPressed: () {
-          onConfirm?.call();
-          Navigator.of(context).pop(true);
+          if (onConfirm != null) {
+            onConfirm!.call();
+          } else {
+            Navigator.pop(context, true);
+          }
         },
+        child: Text(confirmText),
       ),
     ];
 
@@ -80,23 +94,29 @@ class SkadiConfirmationDialog extends StatelessWidget {
   }
 
   Widget _buildAndroidDialog(BuildContext context) {
-    final dangerStyle = TextButton.styleFrom(primary: Colors.red);
+    final dangerStyle = TextButton.styleFrom(foregroundColor: Colors.red);
     final List<Widget> actions = <Widget>[
       TextButton(
         style: !_danger ? dangerStyle : null,
-        child: Text(cancelText),
         onPressed: () {
-          onCancel?.call();
-          Navigator.of(context).pop(false);
+          if (onCancel != null) {
+            onCancel!.call();
+          } else {
+            Navigator.pop(context, false);
+          }
         },
+        child: Text(cancelText),
       ),
       TextButton(
         style: _danger ? dangerStyle : null,
-        child: Text(confirmText),
         onPressed: () {
-          onConfirm?.call();
-          Navigator.of(context).pop(true);
+          if (onConfirm != null) {
+            onConfirm!.call();
+          } else {
+            Navigator.pop(context, true);
+          }
         },
+        child: Text(confirmText),
       ),
     ];
     return AlertDialog(